restructured evaluation code
