+++ src/arc_spice/analysis/prop_models.py
@@ -7,30 +7,7 @@ from typing import Any
 import numpy as np
 from sklearn.linear_model import LinearRegression
 
-from arc_spice.eval import analysis_utils as au
-
-
-def multiplication_prop(
-    results_dict: dict[str, tuple[list[float], list[float]]],
-) -> dict[str, list[float]]:
-    """Compute the naïve approach of multiplying together confidences as though they
-    represent independent probabilities of success.
-
-    Args:
-        results_dict: collated results dict, with structure:
-                        {
-                            'task': (uq vector, error vector)
-                        }
-
-    Returns:
-        multiplied results dict, with same structure but updated uq vectors
-    """
-    previous_vec = np.ones_like(np.array(results_dict["recognition"][0]))
-    mult_res = {}
-    for key, itm in results_dict.items():
-        previous_vec = previous_vec * np.array(itm[0])
-        mult_res[key] = previous_vec
-    return mult_res
+from arc_spice.analysis.utils import brier_score, test_train_split_res
 
 
 def eval_mult_prop(
@@ -50,7 +27,7 @@ def eval_mult_prop(
     mult_res = multiplication_prop(results_dict)
     out = {}
     for key, itm in results_dict.items():
-        out[key] = au.brier_score(mult_res[key], itm[1])
+        out[key] = brier_score(mult_res[key], itm[1])
     return out
 
 
@@ -104,9 +81,52 @@ def fit_uncertainty_model(
     return {"recognition": reg1, "translation": reg2, "classification": reg3}
 
 
-def fitted_uq_model(
-    results_dict: dict[str, tuple[list[float], list[float]]],
-) -> dict[str, tuple[list[float], list[float]]]:
+def multiplication_prop(results_dict, metric_map=None):
+    """Compute the naïve approach of multiplying together confidences as though they
+    represent independent probabilities of success.
+
+    Args:
+        results_dict: collated results dict, with structure:
+                        {
+                            'task': (uq vector, error vector)
+                        }
+
+    Returns:
+        multiplied results dict, with same structure but updated uq vectors
+    """
+    if metric_map is None:
+        metric_map = {
+            "recognition": "mean_entropy",
+            "translation": "weighted_semantic_density",
+            "classification": "mean_predicted_entropy",
+        }
+    # split results and fit models
+    vectors_dict = {
+        "recognition": (
+            1 - np.array(results_dict["recognition"][metric_map["recognition"]]),
+            1 - np.array(results_dict["recognition"]["character_error_rate"]),
+        ),
+        "translation": (
+            np.array(results_dict["translation"][metric_map["translation"]]),
+            np.array(results_dict["translation"]["comet_score"]),
+        ),
+        "classification": (
+            (
+                1
+                - np.array(results_dict["classification"][metric_map["classification"]])
+            ).tolist(),
+            (1 - np.array(results_dict["classification"]["hamming_loss"])).tolist(),
+        ),
+    }
+    previous_vec = np.ones_like(np.array(vectors_dict["recognition"][0]))
+    mult_res = {}
+    for step_key in list(vectors_dict.keys()):
+        previous_vec = previous_vec * np.array(vectors_dict[step_key][0])
+        mult_res[step_key] = previous_vec
+    return mult_res
+
+
+def fitted_uq_model(results_dict, metric_map=None):
     """Fit the uq models using the fit uncertainty models method on a test/train split,
     then populate the data with the test split
 
@@ -120,7 +140,35 @@ def fitted_uq_model(
         test results split with uq propagation from fitted model
     """
     # split results and fit models
-    train_res, test_res = au.test_train_split_res(results_dict)
+    if metric_map is None:
+        metric_map = {
+            "recognition": "mean_entropy",
+            "translation": "weighted_semantic_density",
+            "classification": "mean_predicted_entropy",
+        }
+    # split results and fit models
+    vectors_dict = {
+        "recognition": (
+            1 - np.array(results_dict["recognition"][metric_map["recognition"]]),
+            1 - np.array(results_dict["recognition"]["character_error_rate"]),
+        ),
+        "translation": (
+            np.array(results_dict["translation"][metric_map["translation"]]),
+            np.array(results_dict["translation"]["comet_score"]),
+        ),
+        "classification": (
+            (
+                1
+                - np.array(results_dict["classification"][metric_map["classification"]])
+            ).tolist(),
+            (1 - np.array(results_dict["classification"]["hamming_loss"])).tolist(),
+        ),
+        "celex_ids": (
+            results_dict["classification"]["celex_id"],
+            results_dict["classification"]["celex_id"],
+        ),
+    }
+    train_res, test_res = test_train_split_res(vectors_dict)
     uq_models = fit_uncertainty_model(train_res)
 
     # generated predicted data
@@ -129,21 +177,42 @@ def fitted_uq_model(
     )
     trans_pred = uq_models["translation"].predict(
         np.column_stack(
-            (recog_pred, np.array(test_res["translation"][0]).reshape(-1, 1))
+            (
+                recog_pred,
+                np.array(test_res["translation"][0]).reshape(-1, 1),
+            )
         )
     )
     class_pred = uq_models["classification"].predict(
         np.column_stack(
-            (trans_pred, np.array(test_res["classification"][0]).reshape(-1, 1))
+            (
+                trans_pred,
+                np.array(test_res["classification"][0]).reshape(-1, 1),
+            )
         )
     )
 
     # return collated output
-    return {
-        "recognition": (recog_pred.reshape(1, -1), test_res["recognition"][1]),
-        "translation": (trans_pred.reshape(1, -1), test_res["translation"][1]),
-        "classification": (class_pred.reshape(1, -1), test_res["classification"][1]),
-    }
+    return (
+        {
+            "recognition": (
+                recog_pred.reshape(1, -1).squeeze(),
+                test_res["recognition"][1],
+            ),
+            "translation": (
+                trans_pred.reshape(1, -1).squeeze(),
+                test_res["translation"][1],
+            ),
+            "classification": (
+                class_pred.reshape(1, -1).squeeze(),
+                test_res["classification"][1],
+            ),
+        },
+        {
+            "train_ids": train_res["celex_ids"][0],
+            "test_ids": test_res["celex_ids"][0],
+        },
+    )
 
 
 def eval_lin_models(
@@ -162,7 +231,7 @@ def eval_lin_models(
     # recognition step
     x1 = np.array(test_uq["recognition"][0]).reshape(-1, 1)
     pred1 = lin_uq_models["recognition"].predict(x1)
-    recog_brier = au.brier_score(pred1.reshape(1, -1), test_uq["recognition"][1])
+    recog_brier = brier_score(pred1.reshape(1, -1), test_uq["recognition"][1])
 
     # translation step
     x2 = np.column_stack(
@@ -172,7 +241,7 @@ def eval_lin_models(
         )
     )
     pred2 = lin_uq_models["translation"].predict(x2)
-    trans_brier = au.brier_score(pred2.reshape(1, -1), test_uq["translation"][1])
+    trans_brier = brier_score(pred2.reshape(1, -1), test_uq["translation"][1])
 
     # classification step
     x3 = np.column_stack(
@@ -182,7 +251,7 @@ def eval_lin_models(
         )
     )
     pred3 = lin_uq_models["classification"].predict(x3)
-    class_brier = au.brier_score(pred3.reshape(1, -1), test_uq["classification"][1])
+    class_brier = brier_score(pred3.reshape(1, -1), test_uq["classification"][1])
 
     return {
         "recognition": recog_brier,
